manage read, upload and delete of images based on permissions. Add pre-global error boundary

+++ src/server/queries.ts
@@ -1,6 +1,6 @@
 import "server-only";
 
-import { auth } from "@clerk/nextjs/server";
+import { auth, clerkClient } from "@clerk/nextjs/server";
 import { db } from "~/server/db";
 import { images } from "./db/schema";
 import { and, eq } from "drizzle-orm";
@@ -13,8 +13,13 @@ export async function getMyImages() {
 
   if (!user.userId) throw new Error("Unauthorized");
 
+  const fullUserData = await clerkClient.users.getUser(user.userId);
+
+  if (fullUserData?.privateMetadata?.["can-read"] !== true) {
+    throw new Error("User Does Not Have Read Permissions");
+  }
+
   const images = await db.query.images.findMany({
-    where: (model, { eq }) => eq(model.userId, user.userId),
     orderBy: (model, { desc }) => desc(model.id),
   });
 
@@ -26,13 +31,17 @@ export async function getImage(id: number) {
 
   if (!user.userId) throw new Error("Unauthorized");
 
+  const fullUserData = await clerkClient.users.getUser(user.userId);
+
+  if (fullUserData?.privateMetadata?.["can-read"] !== true) {
+    throw new Error("User Does Not Have Read Permissions");
+  }
+
   const image = await db.query.images.findFirst({
     where: (model, { eq }) => eq(model.id, id),
   });
 
   if (!image) throw new Error("Image not found");
-
-  if (user.userId !== image.userId) throw new Error("Unauthorized");
 
   return image;
 }
@@ -42,14 +51,17 @@ export async function deleteImage(id: number, url: string) {
 
   if (!user.userId) throw new Error("Unauthorized");
 
-  await db
-    .delete(images)
-    .where(and(eq(images.id, id), eq(images.userId, user.userId)));
+  const fullUserData = await clerkClient.users.getUser(user.userId);
+
+  if (fullUserData?.privateMetadata?.["can-delete"] !== true) {
+    throw new Error("User Does Not Have Delete Permissions");
+  }
 
   const urlArray = url.split("/");
-  const deleteKey = urlArray[urlArray.length-1];
+  const deleteKey = urlArray[urlArray.length - 1];
 
   if (typeof deleteKey === "string") {
+    await db.delete(images).where(eq(images.id, id));
     await utapi.deleteFiles(deleteKey);
   }
